Deduplicate directional support powers.
